Expand Search
- Expand searchable fields
- Move filter function externally
