+++ src/services/Utility.js
@@ -1,5 +1,6 @@
 import getPokemonIcon from './functions/getPokemonIcon'
 import getProperName from './functions/getProperName'
+import filterPokemon from './functions/filterPokemon'
 
 class Utility {
   static getPokemonIcon(availableForms, pokemonId, form, evolution, gender, costume, shiny) {
@@ -8,6 +9,10 @@ class Utility {
 
   static getProperName(word) {
     return getProperName(word)
+  }
+
+  static filterPokemon(tempFilters, filters, search) {
+    return filterPokemon(tempFilters, filters, search)
   }
 }
